Modal message and navigate
