+++ src/components/Cart/Cart.scss
@@ -29,6 +29,7 @@
     font-weight: 800;
     line-height: 41px;
     margin-bottom: 24px;
+    margin-top: 24px;
   }
 
   &__items {
